fix(scripts): break file list after invalid

+++ scripts/server-save-action.sh
@@ -18,6 +18,9 @@ done
 
 IFS=$'\n'
 for line in $FILE_LIST; do
+  if [[ ! $line =~ ^clj ]]; then
+    break
+  fi
   file=$(echo "$line" | cut -d':' -f1)
   length=$(echo "$line" | cut -d':' -f2)
   dd if="${ROOT}/worlds/$file" of="${BACKUPS}/${1}/${2}/$file" bs=1 count="$length"
